requirements.txt
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.1.0
six==1.12.0
torch==1.1.0
torchvision==0.3.0
